Rename gaussian_start and add time, also new tests
Renames gaussian_start to gaussian, as it is to be used as the analytical solution
to gaussian starting conditions.
Also renames the test file.
Adds time to the gaussian calc.
Adds a few new tests for the gaussian function

This does not currently match the solutions and is likely incorrect.