Rushed multiple changes
Changes made while working on the dissertaion is the past few weeks

Removes some things from the save files to reduce file size.

Reordered analyse so that is can close the h5 saves to reduce
memory  usage

Modifed plots to be clearer for disseration.

Analyse now creates a storage h5 file, which is a pandas dataframe
with just the importnat bits from
all the runs,
this makes reploting far quicker
but creating the storage file
takes a long time for large runs